refactor: use to user-provided texture, instance and device for rendering
