+++ sparse_strips/vello_hybrid/examples/simple.rs
@@ -7,14 +7,20 @@
 //! It creates a window and continuously renders a simple scene with various shapes.
 //! The rendering pipeline uses CPU-side path processing and GPU-accelerated compositing.
 
-use std::sync::Arc;
+mod common;
 
+use common::{create_vello_renderer, create_winit_window};
 use kurbo::Affine;
 use peniko::{
     color::palette,
     kurbo::{BezPath, Stroke},
 };
-use vello_hybrid::{RenderContext, RenderData, RenderTarget, Renderer, SurfaceTarget};
+use std::sync::Arc;
+use vello_hybrid::{
+    Renderer, Scene,
+    render::RenderParams,
+    util::{RenderContext, RenderSurface},
+};
 use winit::{
     application::ApplicationHandler,
     event::WindowEvent,
@@ -26,8 +32,10 @@ use winit::{
 /// Creates a window and continuously renders a simple scene using the hybrid CPU/GPU renderer.
 fn main() {
     let mut app = SimpleVelloApp {
-        context: None,
-        state: None,
+        context: RenderContext::new(),
+        renderers: vec![],
+        state: RenderState::Suspended(None),
+        scene: Scene::new(900, 600),
     };
 
     let event_loop = EventLoop::new().unwrap();
@@ -36,28 +44,27 @@ fn main() {
         .expect("Couldn't run event loop");
 }
 
-/// State for active rendering
-struct SimpleRenderState {
-    // The window for presenting the rendered content
-    window: Arc<Window>,
-    // The renderer for handling GPU operations
-    renderer: Renderer,
-    // The processed data ready to be sent to the GPU
-    render_data: RenderData,
+#[derive(Debug)]
+enum RenderState<'s> {
+    Active {
+        surface: Box<RenderSurface<'s>>,
+        window: Arc<Window>,
+    },
+    Suspended(Option<Arc<Window>>),
 }
 
-/// Main application state
-struct SimpleVelloApp {
-    // The vello RenderContext which contains rendering state
-    context: Option<RenderContext>,
-    // Active render state (either active or none)
-    state: Option<SimpleRenderState>,
+struct SimpleVelloApp<'s> {
+    context: RenderContext,
+    renderers: Vec<Option<Renderer>>,
+    state: RenderState<'s>,
+    scene: Scene,
 }
 
-impl ApplicationHandler for SimpleVelloApp {
+impl ApplicationHandler for SimpleVelloApp<'_> {
     fn suspended(&mut self, _event_loop: &ActiveEventLoop) {
-        // Release resources when app is suspended
-        self.state = None;
+        if let RenderState::Active { window, .. } = &self.state {
+            self.state = RenderState::Suspended(Some(window.clone()));
+        }
     }
 
     #[allow(
@@ -65,80 +72,111 @@ impl ApplicationHandler for SimpleVelloApp {
         reason = "Width and height are expected to fit within u16 range"
     )]
     fn resumed(&mut self, event_loop: &ActiveEventLoop) {
-        if self.state.is_none() {
-            let window = create_winit_window(event_loop, 900, 600, false);
+        let RenderState::Suspended(cached_window) = &mut self.state else {
+            return;
+        };
 
-            let mut context = RenderContext::new(
-                window.inner_size().width as u16,
-                window.inner_size().height as u16,
-            );
-            draw_simple_scene(&mut context);
-            let render_data = context.prepare_render_data();
-            self.context = Some(context);
+        let window = cached_window.take().unwrap_or_else(|| {
+            create_winit_window(
+                event_loop,
+                self.scene.width() as u32,
+                self.scene.height() as u32,
+                true,
+            )
+        });
 
-            let mut renderer = pollster::block_on(async {
-                let window_size = window.inner_size();
-                Renderer::new(RenderTarget::Surface {
-                    target: Arc::clone(&window) as Arc<dyn SurfaceTarget>,
-                    width: window_size.width,
-                    height: window_size.height,
-                })
-                .await
-            });
+        let size = window.inner_size();
+        let surface = pollster::block_on(self.context.create_surface(
+            window.clone(),
+            size.width,
+            size.height,
+            wgpu::PresentMode::AutoVsync,
+            wgpu::TextureFormat::Bgra8Unorm,
+        ));
 
-            renderer.prepare(&render_data);
+        self.renderers
+            .resize_with(self.context.devices.len(), || None);
+        self.renderers[surface.dev_id]
+            .get_or_insert_with(|| create_vello_renderer(&self.context, &surface));
 
-            window.set_visible(true);
-            window.focus_window();
-            window.request_redraw();
-
-            self.state = Some(SimpleRenderState {
-                window,
-                renderer,
-                render_data,
-            });
-        }
+        self.state = RenderState::Active {
+            surface: Box::new(surface),
+            window,
+        };
     }
 
     fn window_event(
         &mut self,
         event_loop: &ActiveEventLoop,
-        _window_id: WindowId,
+        window_id: WindowId,
         event: WindowEvent,
     ) {
-        if let Some(state) = &self.state {
-            match event {
-                WindowEvent::RedrawRequested => {
-                    state.renderer.render_to_surface(&state.render_data);
-                    state.window.request_redraw();
-                }
-                WindowEvent::CloseRequested => {
-                    event_loop.exit();
-                }
-                _ => {}
+        let surface = match &mut self.state {
+            RenderState::Active { surface, window } if window.id() == window_id => surface,
+            _ => return,
+        };
+
+        match event {
+            WindowEvent::CloseRequested => event_loop.exit(),
+            WindowEvent::Resized(size) => {
+                self.context
+                    .resize_surface(surface, size.width, size.height);
             }
+            WindowEvent::RedrawRequested => {
+                self.scene.reset();
+                draw_simple_scene(&mut self.scene);
+
+                let width = surface.config.width;
+                let height = surface.config.height;
+                let device_handle = &self.context.devices[surface.dev_id];
+
+                self.renderers[surface.dev_id]
+                    .as_mut()
+                    .unwrap()
+                    .render_to_texture(
+                        &device_handle.device,
+                        &device_handle.queue,
+                        &self.scene,
+                        &surface.target_view,
+                        &RenderParams {
+                            base_color: palette::css::BLACK,
+                            width,
+                            height,
+                            strip_height: 4,
+                        },
+                    );
+
+                let surface_texture = surface
+                    .surface
+                    .get_current_texture()
+                    .expect("failed to get surface texture");
+
+                let mut encoder =
+                    device_handle
+                        .device
+                        .create_command_encoder(&wgpu::CommandEncoderDescriptor {
+                            label: Some("Surface Blit"),
+                        });
+                surface.blitter.copy(
+                    &device_handle.device,
+                    &mut encoder,
+                    &surface.target_view,
+                    &surface_texture
+                        .texture
+                        .create_view(&wgpu::TextureViewDescriptor::default()),
+                );
+                device_handle.queue.submit([encoder.finish()]);
+                surface_texture.present();
+
+                device_handle.device.poll(wgpu::Maintain::Poll);
+            }
+            _ => {}
         }
     }
 }
 
-/// Helper function that creates a Winit window and returns it (wrapped in an Arc for sharing)
-fn create_winit_window(
-    event_loop: &ActiveEventLoop,
-    width: u32,
-    height: u32,
-    initially_visible: bool, // Whether the window should be visible initially or made visible later
-) -> Arc<Window> {
-    let attr = Window::default_attributes()
-        .with_inner_size(winit::dpi::PhysicalSize::new(width, height))
-        .with_resizable(false)
-        .with_title("Vello Simple Renderer")
-        .with_visible(initially_visible)
-        .with_active(true);
-    Arc::new(event_loop.create_window(attr).unwrap())
-}
-
 /// Draws a simple scene with shapes
-fn draw_simple_scene(ctx: &mut RenderContext) {
+fn draw_simple_scene(ctx: &mut Scene) {
     let mut path = BezPath::new();
     path.move_to((10.0, 10.0));
     path.line_to((180.0, 20.0));
